interview_data_update/requirements.txt
psycopg2-binary>=2.9.0
python-dateutil>=2.8.0
boto3>=1.26.0
botocore>=1.29.0






